% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COpts}
\alias{COpts}
\title{CO point data}
\format{
An \code{sfc_POINT} object with 100 points
}
\usage{
COpts
}
\description{
A set of 100 randomly-selected point locations from within Colorado
}
\keyword{datasets}
