% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomaskr.R
\docType{package}
\name{geomaskr-package}
\alias{geomaskr-package}
\title{Geomask Locations in R}
\description{
Provides functions to conduct geomasking of coordinate locations. This includes using a "k-anonymity" approach to selecting a radius for location perturbation.
}
