% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sampling.R
\name{sample_circle}
\alias{sample_circle}
\alias{sample_circle_boundedbox}
\title{sample_circle}
\usage{
sample_circle(x, y, r, r0 = 0)

sample_circle_boundedbox(
  x,
  y,
  r,
  r0 = 0,
  xmin = 0,
  xmax = 1,
  ymin = 0,
  ymax = 1,
  max_redraw = 100
)
}
\arguments{
\item{x}{First coordinate for the location(s) to be masked}

\item{y}{Second coordinate for the location(s) to be mased}

\item{r}{Radius of masking region}

\item{r0}{Minimum radius for masking. Values greater than zero lead to "donut geomasking"}

\item{xmin}{lower limit of bounding box}

\item{xmax}{upper limit of bounding box}

\item{ymin}{lower limit of bounding box}

\item{ymax}{upper limit of bounding box}

\item{max_redraw}{number of times to draw again, if first draw is outside the bounding box}
}
\value{
Data frame containing `x` and `y` coordinates of the perturbed locations
}
\description{
sample_circle

Title
}
\examples{
sample_circle(x=0, y=0, r=1, r0=0)
}
