% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sampling.R
\name{st_sample_radius}
\alias{st_sample_radius}
\alias{st_sample_radius_bounded}
\alias{st_sample_radius_bounded_set}
\title{Title
Sample Points within a Radius}
\usage{
st_sample_radius(pt, radius = 100, region)

st_sample_radius_bounded(
  pt,
  radius = 100,
  region,
  minradius = 0,
  maxretry = 1000,
  return_dist = TRUE
)

st_sample_radius_bounded_set(
  pts,
  radius = 100,
  region,
  minradius = 0,
  k_anon = NULL,
  pop_density = NULL,
  verbose = FALSE,
  maxretry = 1000
)
}
\arguments{
\item{pt}{Point location to perturb. Assumed to be an}

\item{radius}{Radius of circle to sample within}

\item{region}{A \code{sfc} object used to bound the perturbed locations.}

\item{minradius}{Minimum radius}

\item{maxretry}{Number of times to retry point selection if selected point is either inside minimum radius or not in bounding region}

\item{return_dist}{Logical indicator of whether or not to return the distance of new point from original point}

\item{pts}{Collection of points to resample location for}

\item{k_anon}{K-anonymity value. Used for calculating the radius. Not used unless k_anon and pop_density are both provided.}

\item{pop_density}{Population density, assumed to be in people per km^2. Used for calculating the radius. Not used unless k_anon and pop_density are both provided.}

\item{verbose}{Logical for printing the calculated radii to output.}
}
\value{
\code{sfc} object containg the new point
}
\description{
Randomly samples points within a given radius using  functions from `sf` and `spatstat`.
}
\details{
New point is transformed to CRS of `pt`
}
