% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sampling.R
\name{calc_radius}
\alias{calc_radius}
\alias{calc_k}
\title{calc_radius}
\usage{
calc_radius(k_anon, density, units_scale = 1)

calc_k(radius, density, units_scale = 1)
}
\arguments{
\item{k_anon}{K-anonymity metric}
}
\description{
Calculates radius for a specified k-anonymity and population density.
}
